webpersonal/requirements.txt
pillow
pyserial
SpeechRecognition==3.10.0
pyttsx3
newsapi-python
numpy
opencv-python
mediapipe
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
idna==3.4
psycopg2-binary==2.9.6
requests==2.31.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.3

